Update jquery.maskMoney.js
a little improvement, which allows to set the currency symbol at the end of the mask Eg 10 000 USD, 122 € etc
All you have to do is set an extra parameter: symbolPosition: 'after' (‘before’ by default).

Also I trigger the blur event after setting a mask to a field with
value already set. This aplly the mask to it automaticly on init.